ImageProsessing/p_ImageRestoration/ILL_Posed/Deblur_DL/Black-box image deblurring and defiltering/requirements.txt
numpy
scipy
matplotlib
scikit-image
Pillow
